Refactoring Common GenericServices
