+++ src/GSWCloudApp.Common/ServiceGenerics/DependencyInjection.cs
@@ -1,4 +1,5 @@
 ﻿using GSWCloudApp.Common.ServiceGenerics.Services;
+using GSWCloudApp.Common.ServiceGenerics.Services.Interfaces;
 using Microsoft.Extensions.DependencyInjection;
 
 namespace GSWCloudApp.Common.Services;
@@ -11,5 +12,8 @@ public static class DependencyInjection
     /// <param name="services">The service collection to add the generic services to.</param>
     /// <returns>The updated service collection.</returns>
     public static IServiceCollection ConfigureGenericServices(this IServiceCollection services)
-        => services.AddTransient<IGenericService, GenericService>();
+        => services
+            .AddTransient<IGenericService, GenericService>()
+            //.AddTransient<ICachedGenericService, CachedGenericService>()
+            ;
 }
